feat(bigint): auto-promotion closes DIFF-003 documented gaps (PMAT-036)
Converts the 20 documented promotion gaps in the differential-exec
gate from panics into successful BigInt-equivalent outputs.

Before:
  XPILE-DIFF-003: 20 overflow-phase checks across 2 fixture(s) —
  20 documented promotion gaps, 0 promoted-and-agreed.

After:
  XPILE-DIFF-003: 20 overflow-phase checks across 2 fixture(s) —
  0 documented promotion gaps, 20 promoted-and-agreed.

Mechanism — exercises existing PMAT-013 / PMAT-025 infrastructure
on the overflow-prone fixtures, plus three small targeted fixes:

1. factorial.py + countdown.py annotated `-> BigInt`. PMAT-013's
   implicit-promotion lifts `n: int` to BigInt and every int
   literal in the body to `xpile_bigint::BigInt::from(...)`, so
   the whole function runs in BigInt mode end-to-end and never
   overflows for n in [21, 30].

2. depyler-frontend extends BigInt-mode propagation to for-range
   loop-target bindings. Before: `for i in range(n, 0, -1)`
   lowered to `let mut i: i64 = n` even with `n: BigInt` — a
   type error. Now follows `ctx.fn_return_type`, so BigInt-mode
   functions get BigInt loop variables. Required for the
   countdown.py iterative case to compile.

3. depyler-frontend skips `from __future__ import annotations`
   as a no-op preamble. Required for CPython to `exec` the
   fixture without `NameError: BigInt` (xpile's metadata-only
   type alias for Python's unbounded int).

4. diff_exec.rs dual-mode build pipeline. When the transpile
   output uses xpile_bigint::BigInt, the runner materialises a
   one-shot Cargo project that depends on the in-workspace
   xpile-bigint crate (path dep) so the produced binary has the
   real num_bigint::BigInt + Display. Non-BigInt fixtures keep
   the existing standalone-rustc fast path. Cargo build pins
   --target-dir so the binary lands at a predictable path
   regardless of any global CARGO_TARGET_DIR / .cargo/config.toml.

E2E test updates: 3 transpile_e2e tests that hard-asserted i64
emission for factorial/countdown now assert BigInt emission.
Drivers use inline `mod xpile_bigint { ... }` shims matching the
existing PMAT-013 BigInt fixture tests.

Architectural payoff: this PR proves the §27 type lattice handles
dynamic size escalation through a complete fixture lifecycle —
frontend lowering, codegen, and differential-exec all participate
in the BigInt-mode path. The 20→0 documented-gap flip is the
user-visible metric.

Gates: cargo fmt, clippy -D warnings, cargo test --workspace,
cargo deny check advisories, pv lint contracts/ — all green.

(Refs PMAT-036)

+++ crates/xpile/tests/fixtures/countdown.py
@@ -1,11 +1,25 @@
-# Negative-step for-range fixture (PMAT-008).
+# Negative-step for-range fixture (PMAT-008) + BigInt promotion
+# (PMAT-036).
 #
-# `factorial_iter(n)` = n! computed by counting down with `range(n, 0, -1)`.
+# `factorial_iter(n)` = n! computed by counting down with
+# `range(n, 0, -1)`. Annotating `-> BigInt` triggers PMAT-013's
+# implicit promotion: every `int` param + literal in the body
+# lifts to BigInt, so the iterative multiplication never overflows.
+#
 # Exercises:
 #   - negative step → cond is `i > stop`, not `i < stop`
-#   - the tail `i = i + (-1)` still uses checked_add (which handles
-#     adding negative i64s natively)
-def factorial_iter(n: int) -> int:
+#   - the tail `i = i + (-1)` uses BigInt arithmetic (the literal
+#     `-1` lifts to BigInt under PMAT-013's literal-promotion rule)
+#   - BigInt mode through a `for-range` loop end-to-end (PMAT-036's
+#     loop-target type fix in depyler-frontend).
+#
+# `from __future__ import annotations` defers annotation evaluation
+# so `BigInt` (xpile's metadata-only alias for Python's unbounded
+# `int`) doesn't NameError at function-def time when CPython
+# imports / execs the file in the diff_exec gate.
+from __future__ import annotations
+
+def factorial_iter(n: int) -> BigInt:
     acc = 1
     for i in range(n, 0, -1):
         acc = acc * i
